+++ source/config/configFlags.py
@@ -225,3 +225,23 @@ class AddonsAutomaticUpdate(DisplayStringStrEnum):
 			# Translators: This is a label for the automatic update behaviour for add-ons.
 			self.DISABLED: _("Disabled"),
 		}
+
+
+class ResetNewAddons(DisplayStringStrEnum):
+	MONTHLY = "monthly"
+	WEEKLY = "weekly"
+	STARTUP = "startup"
+
+	@property
+	def _displayStringLabels(self):
+		return {
+			# Translators: This is a label for the reset new add-ons behavior.
+			# It will determine the period for an add-on to be considered as new.
+			self.MONTHLY: _("Monthly"),
+			# Translators: This is a label for the reset new add-ons behavior.
+			# It will determine the period for an add-on to be considered as new.
+			self.WEEKLY: _("Weekly"),
+			# Translators: This is a label for the reset new add-ons behavior.
+			# It will determine the period for an add-on to be considered as new.
+			self.STARTUP: _("At startup"),
+		}